requirements.txt
Flask==2.3.2
opencv-python==4.9.0.80
numpy==1.24.4
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.3.0
matplotlib==3.7.2
pillow==10.2.0
werkzeug==2.3.6
gunicorn==21.2.0
